extend the component store api
