+++ packages/packages/core/editor-components/src/store/store.ts
@@ -1,176 +1,14 @@
-import { type V1Document } from '@elementor/editor-documents';
-import {
-	__createSelector as createSelector,
-	__createSlice as createSlice,
-	__useSelector as useSelector,
-	type PayloadAction,
-	type SliceState,
-} from '@elementor/store';
+import { __createSelector as createSelector, __useSelector as useSelector, type SliceState } from '@elementor/store';
 
-import {
-	type Component,
-	type ComponentId,
-	type OverridableProps,
-	type PublishedComponent,
-	type StylesDefinition,
-	type UnpublishedComponent,
-} from '../types';
-import { loadComponents } from './thunks';
-
-type GetComponentResponse = PublishedComponent[];
-
-type Status = 'idle' | 'pending' | 'error';
-
-export type SanitizeAttributes = 'overridableProps';
-
-type ComponentsState = {
-	data: PublishedComponent[];
-	unpublishedData: UnpublishedComponent[];
-	loadStatus: Status;
-	styles: StylesDefinition;
-	createdThisSession: Component[ 'uid' ][];
-	archivedThisSession: ComponentId[];
-	path: ComponentsPathItem[];
-	currentComponentId: V1Document[ 'id' ] | null;
-	updatedComponentNames: Record< number, string >;
-
-	// We use this map to flag any sanitized attribute of a given component
-	// This map currently resets in response to the `editor/documents/open` command
-	sanitized: Record< ComponentId, Partial< Record< SanitizeAttributes, boolean > > >;
-};
+import type { ComponentId, OverridableProps, PublishedComponent, UnpublishedComponent } from '../types';
+import { type slice } from './extensible-slice';
+import { type SanitizeAttributes, SLICE_NAME } from './store-types';
 
 export type ComponentsSlice = SliceState< typeof slice >;
 
-export type ComponentsPathItem = {
-	instanceId?: string;
-	instanceTitle?: string;
-	componentId: V1Document[ 'id' ];
-};
-
-export const initialState: ComponentsState = {
-	data: [],
-	unpublishedData: [],
-	loadStatus: 'idle',
-	styles: {},
-	createdThisSession: [],
-	archivedThisSession: [],
-	path: [],
-	currentComponentId: null,
-	updatedComponentNames: {},
-	sanitized: {},
-};
-
-export const SLICE_NAME = 'components';
-
-export const slice = createSlice( {
-	name: SLICE_NAME,
-	initialState,
-	reducers: {
-		add: ( state, { payload }: PayloadAction< PublishedComponent | PublishedComponent[] > ) => {
-			if ( Array.isArray( payload ) ) {
-				state.data = [ ...payload, ...state.data ];
-			} else {
-				state.data.unshift( payload );
-			}
-		},
-		load: ( state, { payload }: PayloadAction< PublishedComponent[] > ) => {
-			state.data = payload;
-		},
-		addUnpublished: ( state, { payload }: PayloadAction< UnpublishedComponent > ) => {
-			state.unpublishedData.unshift( payload );
-		},
-		removeUnpublished: ( state, { payload }: PayloadAction< string | string[] > ) => {
-			const uidsToRemove = Array.isArray( payload ) ? payload : [ payload ];
-			state.unpublishedData = state.unpublishedData.filter(
-				( component ) => ! uidsToRemove.includes( component.uid )
-			);
-		},
-		resetUnpublished: ( state ) => {
-			state.unpublishedData = [];
-		},
-		removeStyles( state, { payload }: PayloadAction< { id: ComponentId } > ) {
-			const { [ payload.id ]: _, ...rest } = state.styles;
-
-			state.styles = rest;
-		},
-		addStyles: ( state, { payload } ) => {
-			state.styles = { ...state.styles, ...payload };
-		},
-		addCreatedThisSession: ( state, { payload }: PayloadAction< string > ) => {
-			state.createdThisSession.push( payload );
-		},
-		removeCreatedThisSession: ( state, { payload }: PayloadAction< string > ) => {
-			state.createdThisSession = state.createdThisSession.filter( ( uid ) => uid !== payload );
-		},
-		archive: ( state, { payload }: PayloadAction< number > ) => {
-			const component = state.data.find( ( comp ) => comp.id === payload );
-
-			if ( component ) {
-				component.isArchived = true;
-				state.archivedThisSession.push( payload );
-			}
-		},
-		setCurrentComponentId: ( state, { payload }: PayloadAction< V1Document[ 'id' ] | null > ) => {
-			state.currentComponentId = payload;
-		},
-		setPath: ( state, { payload }: PayloadAction< ComponentsPathItem[] > ) => {
-			state.path = payload;
-		},
-		setOverridableProps: (
-			state,
-			{ payload }: PayloadAction< { componentId: ComponentId; overridableProps: OverridableProps } >
-		) => {
-			const component = state.data.find( ( comp ) => comp.id === payload.componentId );
-
-			if ( ! component ) {
-				return;
-			}
-
-			component.overridableProps = payload.overridableProps;
-		},
-		rename: ( state, { payload }: PayloadAction< { componentUid: string; name: string } > ) => {
-			const component = state.data.find( ( comp ) => comp.uid === payload.componentUid );
-
-			if ( ! component ) {
-				return;
-			}
-			if ( component.id ) {
-				state.updatedComponentNames[ component.id ] = payload.name;
-			}
-			component.name = payload.name;
-		},
-		cleanUpdatedComponentNames: ( state ) => {
-			state.updatedComponentNames = {};
-		},
-		updateComponentSanitizedAttribute: (
-			state,
-			{
-				payload: { componentId, attribute },
-			}: PayloadAction< { componentId: ComponentId; attribute: SanitizeAttributes } >
-		) => {
-			if ( ! state.sanitized[ componentId ] ) {
-				state.sanitized[ componentId ] = {};
-			}
-
-			state.sanitized[ componentId ][ attribute ] = true;
-		},
-		resetSanitizedComponents: ( state ) => {
-			state.sanitized = {};
-		},
-	},
-	extraReducers: ( builder ) => {
-		builder.addCase( loadComponents.fulfilled, ( state, { payload }: PayloadAction< GetComponentResponse > ) => {
-			state.data = payload;
-			state.loadStatus = 'idle';
-		} );
-		builder.addCase( loadComponents.pending, ( state ) => {
-			state.loadStatus = 'pending';
-		} );
-		builder.addCase( loadComponents.rejected, ( state ) => {
-			state.loadStatus = 'error';
-		} );
-	},
-} );
+export { slice, registerComponentsReducer, createComponentsAction, __resetExtraReducers } from './extensible-slice';
+export type { ComponentsState, ComponentsPathItem, SanitizeAttributes } from './store-types';
+export { initialState, SLICE_NAME } from './store-types';
 
 export const selectData = ( state: ComponentsSlice ) => state[ SLICE_NAME ].data;
 export const selectArchivedThisSession = ( state: ComponentsSlice ) => state[ SLICE_NAME ].archivedThisSession;
